Remove ability to create custom roles from ConfigurationRole
