+++ subprojects/core/src/main/java/org/gradle/api/internal/artifacts/configurations/ConfigurationRole.java
@@ -16,11 +16,7 @@
 
 package org.gradle.api.internal.artifacts.configurations;
 
-import org.gradle.internal.deprecation.DeprecationLogger;
-
-import javax.annotation.Nullable;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 
 /**
@@ -51,58 +47,10 @@ public interface ConfigurationRole {
     }
 
     /**
-     * Attempts to locate a pre-defined role allowing the given usage in {@link ConfigurationRoles} and return it;
-     * if such a roles does not exist, creates a custom anonymous implementation and returns it instead.
-     *
-     * @param name the name to use when creating a role that wasn't found
-     * @param consumable whether this role is consumable
-     * @param resolvable whether this role is resolvable
-     * @param declarableAgainst whether this role is declarable against
-     * @param consumptionDeprecated whether this role is deprecated for consumption
-     * @param resolutionDeprecated whether this role is deprecated for resolution
-     * @param declarationAgainstDeprecated whether this role is deprecated for declaration against
-     * @param description a custom description to use when creating a role that wasn't found; or {@code null} to use the default description generated by {@link RoleDescriber}
-     * @param warnOnCustomRole {@code true} to warn if this call results in the creation of a custom role; {@code false} otherwise
-     * @return a role with matching usage characteristics
-     */
-    static ConfigurationRole forUsage(String name, boolean consumable, boolean resolvable, boolean declarableAgainst, boolean consumptionDeprecated, boolean resolutionDeprecated, boolean declarationAgainstDeprecated, @Nullable String description, boolean warnOnCustomRole) {
-        if ((!consumable && consumptionDeprecated) || (!resolvable && resolutionDeprecated) || (!declarableAgainst && declarationAgainstDeprecated)) {
-            throw new IllegalArgumentException("Cannot create a role that deprecates a usage that is not allowed");
-        }
-
-        // Built in configuration roles are all deprecations = false, so attempt to match on allowed usage, then check if the deprecations happen to be all false on the match
-        ConfigurationRole result = ConfigurationRoles.byUsage(consumable, resolvable, declarableAgainst)
-                .map(ConfigurationRole.class::cast)
-                .filter(role -> role.isConsumptionDeprecated() == consumptionDeprecated && role.isResolutionDeprecated() == resolutionDeprecated && role.isDeclarationAgainstDeprecated() == declarationAgainstDeprecated)
-                .orElse(new CustomConfigurationRole(name, consumable, resolvable, declarableAgainst, consumptionDeprecated, resolutionDeprecated, declarationAgainstDeprecated, description));
-
-        //noinspection SuspiciousMethodCalls
-        if (warnOnCustomRole && !Arrays.asList(ConfigurationRoles.values()).contains(result)) {
-            DeprecationLogger.deprecateBehaviour("Custom configuration roles are deprecated.")
-                    .withAdvice("Use one of the standard roles defined in ConfigurationRoles instead.")
-                    .willBeRemovedInGradle9()
-                    .withUpgradeGuideSection(8, "custom_configuration_roles")
-                    .nagUser();
-        }
-
-        return result;
-    }
-
-    static ConfigurationRole forUsage(boolean consumable, boolean resolvable, boolean declarableAgainst, boolean consumptionDeprecated, boolean resolutionDeprecated, boolean declarationAgainstDeprecated) {
-        return forUsage(RoleDescriber.DEFAULT_CUSTOM_ROLE_NAME, consumable, resolvable, declarableAgainst, consumptionDeprecated, resolutionDeprecated, declarationAgainstDeprecated);
-    }
-
-    static ConfigurationRole forUsage(String name, boolean consumable, boolean resolvable, boolean declarableAgainst, boolean consumptionDeprecated, boolean resolutionDeprecated, boolean declarationAgainstDeprecated) {
-        return forUsage(name, consumable, resolvable, declarableAgainst, consumptionDeprecated, resolutionDeprecated, declarationAgainstDeprecated, null, false);
-    }
-
-    /**
      * This static util class hides methods internal to the {@code default} methods of {@link ConfigurationRole} which
      * can be used to build a human-readable description of the usage a role allows.
      */
     abstract class RoleDescriber {
-        private static final String DEFAULT_CUSTOM_ROLE_NAME = "Custom Role";
-
         private static final String CONSUMABLE = "Consumable - this configuration can be selected by another project as a dependency";
         private static final String RESOLVABLE = "Resolvable - this configuration can be resolved by this project to a set of files";
         private static final String DECLARABLE_AGAINST = "Declarable Against - this configuration can have dependencies added to it";
@@ -137,76 +85,6 @@ public interface ConfigurationRole {
 
         private static String describeDeprecation(boolean deprecated) {
             return deprecated ? " " + IS_DEPRECATED : "";
-        }
-    }
-
-    /**
-     * A custom implementation of {@link ConfigurationRole} that allows for non-standard combinations of usage characteristics and deprecations.
-     */
-    final class CustomConfigurationRole implements ConfigurationRole {
-        private final String name;
-        private final boolean consumable;
-        private final boolean resolvable;
-        private final boolean declarableAgainst;
-        private final boolean consumptionDeprecated;
-        private final boolean resolutionDeprecated;
-        private final boolean declarationAgainstDeprecated;
-        @Nullable
-        private final String description;
-
-        private CustomConfigurationRole(String name, boolean consumable, boolean resolvable, boolean declarableAgainst, boolean consumptionDeprecated, boolean resolutionDeprecated, boolean declarationAgainstDeprecated, @Nullable String description) {
-            this.name = name;
-            this.consumable = consumable;
-            this.resolvable = resolvable;
-            this.declarableAgainst = declarableAgainst;
-            this.consumptionDeprecated = consumptionDeprecated;
-            this.resolutionDeprecated = resolutionDeprecated;
-            this.declarationAgainstDeprecated = declarationAgainstDeprecated;
-            this.description = description;
-        }
-
-        @Override
-        public String getName() {
-            return name;
-        }
-
-        @Override
-        public boolean isConsumable() {
-            return consumable;
-        }
-
-        @Override
-        public boolean isResolvable() {
-            return resolvable;
-        }
-
-        @Override
-        public boolean isDeclarableAgainst() {
-            return declarableAgainst;
-        }
-
-        @Override
-        public boolean isConsumptionDeprecated() {
-            return consumptionDeprecated;
-        }
-
-        @Override
-        public boolean isResolutionDeprecated() {
-            return resolutionDeprecated;
-        }
-
-        @Override
-        public boolean isDeclarationAgainstDeprecated() {
-            return declarationAgainstDeprecated;
-        }
-
-        @Override
-        public String describeUsage() {
-            if (description != null) {
-                return description;
-            } else {
-                return RoleDescriber.describeRole(this);
-            }
         }
     }
 }
