+++ subprojects/dependency-management/src/test/groovy/org/gradle/api/internal/artifacts/configurations/ConfigurationRoleSpec.groovy
@@ -21,73 +21,6 @@ import spock.lang.Specification
 import static org.gradle.api.internal.artifacts.configurations.ConfigurationRole.RoleDescriber.*
 
 class ConfigurationRoleSpec extends Specification {
-    def "can find predefined role #role"() {
-        expect:
-        ConfigurationRole.forUsage(consumable, resolvable, declarableAgainst, consumptionDeprecated, resolutionDeprecated, declarationAgainstDeprecated) == role
-
-        where:
-        consumable  | resolvable    | declarableAgainst | consumptionDeprecated | resolutionDeprecated  | declarationAgainstDeprecated  || role
-        true        | true          | true              | false                 | false                 | false                         || ConfigurationRoles.LEGACY
-        true        | false         | false             | false                 | false                 | false                         || ConfigurationRoles.INTENDED_CONSUMABLE
-        false       | true          | false             | false                 | false                 | false                         || ConfigurationRoles.INTENDED_RESOLVABLE
-        false       | true          | true              | false                 | false                 | false                         || ConfigurationRoles.INTENDED_RESOLVABLE_BUCKET
-        false       | false         | true              | false                 | false                 | false                         || ConfigurationRoles.INTENDED_BUCKET
-    }
-
-    def "can create role for unknown usage combinations consumable=#consumable, resolvable=#resolvable, declarableAgainst=#declarableAgainst, consumptionDeprecated=#consumptionDeprecated, resolutionDeprecated=#resolutionDeprecated, declarationAgainstDeprecated=#declarationAgainstDeprecated"() {
-        expect:
-        ConfigurationRole role = ConfigurationRole.forUsage(consumable, resolvable, declarableAgainst, consumptionDeprecated, resolutionDeprecated, declarationAgainstDeprecated)
-        role.isConsumable() == consumable
-        role.isResolvable() == resolvable
-        role.isDeclarableAgainst() == declarableAgainst
-        role.isConsumptionDeprecated() == consumptionDeprecated
-        role.isResolutionDeprecated() == resolutionDeprecated
-        role.isDeclarationAgainstDeprecated() == declarationAgainstDeprecated
-
-        where:
-        consumable  | resolvable    | declarableAgainst | consumptionDeprecated | resolutionDeprecated  | declarationAgainstDeprecated
-        true        | true          | false             | false                 | false                 | false
-        false       | true          | true              | false                 | true                  | false
-        true        | true          | true              | true                  | true                  | true
-        true        | false         | true              | true                  | false                 | true
-        true        | true          | true              | false                 | true                  | false
-    }
-
-    def "custom role is named correctly"() {
-        when:
-        def customRole = ConfigurationRole.forUsage(consumable, resolvable, declarableAgainst, consumptionDeprecated, resolutionDeprecated, declarationAgainstDeprecated)
-
-        then:
-        customRole !in ConfigurationRoles.values()
-        customRole.name == DEFAULT_CUSTOM_ROLE_NAME
-
-        where:
-        consumable  | resolvable    | declarableAgainst | consumptionDeprecated | resolutionDeprecated  | declarationAgainstDeprecated
-        true        | true          | false             | false                 | false                 | false
-        false       | true          | true              | false                 | true                  | false
-        true        | true          | true              | true                  | true                  | true
-        true        | false         | true              | true                  | false                 | true
-        true        | true          | true              | false                 | true                  | false
-    }
-
-    def "custom role can be given custom description"() {
-        when:
-        def customRole = ConfigurationRole.forUsage('custom', consumable, resolvable, declarableAgainst, consumptionDeprecated, resolutionDeprecated, declarationAgainstDeprecated, 'custom description', false)
-
-        then:
-        customRole !in ConfigurationRoles.values()
-        customRole.name == 'custom'
-        customRole.describeUsage() == 'custom description'
-
-        where:
-        consumable  | resolvable    | declarableAgainst | consumptionDeprecated | resolutionDeprecated  | declarationAgainstDeprecated
-        true        | true          | false             | false                 | false                 | false
-        false       | true          | true              | false                 | true                  | false
-        true        | true          | true              | true                  | true                  | true
-        true        | false         | true              | true                  | false                 | true
-        true        | true          | true              | false                 | true                  | false
-    }
-
     def "roles can describe themselves #role"() {
         expect:
         assertDescriptionContains(role, usages)
@@ -99,21 +32,6 @@ class ConfigurationRoleSpec extends Specification {
         ConfigurationRoles.INTENDED_RESOLVABLE          || [RESOLVABLE]
         ConfigurationRoles.INTENDED_RESOLVABLE_BUCKET   || [RESOLVABLE, DECLARABLE_AGAINST]
         ConfigurationRoles.INTENDED_BUCKET              || [DECLARABLE_AGAINST]
-    }
-
-    def "custom role can't deprecate what it doesn't allow"() {
-        when:
-        ConfigurationRole.forUsage(consumable, resolvable, declarableAgainst, consumptionDeprecated, resolutionDeprecated, declarationAgainstDeprecated)
-
-        then:
-        def e = thrown(IllegalArgumentException)
-        e.message == 'Cannot create a role that deprecates a usage that is not allowed'
-
-        where:
-        consumable  | resolvable    | declarableAgainst | consumptionDeprecated | resolutionDeprecated  | declarationAgainstDeprecated
-        false       | false         | false             | true                  | false                 | false
-        false       | false         | false             | false                 | true                  | false
-        false       | false         | false             | false                 | false                 | true
     }
 
     private void assertDescriptionContains(ConfigurationRole role, List<String> usages) {
